HDDS-3684. Add tests for replication annotation
